requirements.txt
requests
beautifulsoup4
mecab-python3
sklearn